move session_manager into server
