+++ src/connection/server/mod.rs
@@ -21,6 +21,8 @@
 
 mod database;
 mod session;
+pub mod session_manager;
 mod transaction;
 
+pub(crate) use self::session_manager::SessionManager;
 pub use self::{database::Database, session::Session, transaction::Transaction};
